[CHORE] Avatar 도메인 객체를 user 패키지로 이전
