+++ src/main/java/org/websoso/WSSServer/dto/block/BlockGetResponse.java
@@ -1,8 +1,7 @@
 package org.websoso.WSSServer.dto.block;
 
-import org.websoso.WSSServer.domain.Avatar;
-import org.websoso.WSSServer.domain.AvatarProfile;
-import org.websoso.WSSServer.domain.Block;
+import org.websoso.WSSServer.user.domain.AvatarProfile;
+import org.websoso.WSSServer.user.domain.Block;
 import org.websoso.WSSServer.user.domain.User;
 
 public record BlockGetResponse(
