+++ src/main/java/org/websoso/WSSServer/dto/popularNovel/PopularNovelGetResponse.java
@@ -1,7 +1,6 @@
 package org.websoso.WSSServer.dto.popularNovel;
 
-import org.websoso.WSSServer.domain.Avatar;
-import org.websoso.WSSServer.domain.AvatarProfile;
+import org.websoso.WSSServer.user.domain.AvatarProfile;
 import org.websoso.WSSServer.feed.domain.Feed;
 import org.websoso.WSSServer.novel.domain.Novel;
