+++ src/main/java/org/websoso/WSSServer/repository/AvatarProfileRepository.java
@@ -3,7 +3,7 @@ package org.websoso.WSSServer.repository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.websoso.WSSServer.domain.AvatarProfile;
+import org.websoso.WSSServer.user.domain.AvatarProfile;
 
 @Repository
 public interface AvatarProfileRepository extends JpaRepository<AvatarProfile, Long> {
