+++ src/main/java/org/websoso/WSSServer/repository/BlockRepository.java
@@ -3,7 +3,7 @@ package org.websoso.WSSServer.repository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.websoso.WSSServer.domain.Block;
+import org.websoso.WSSServer.user.domain.Block;
 
 @Repository
 public interface BlockRepository extends JpaRepository<Block, Long> {
