+++ src/main/java/org/websoso/WSSServer/service/AvatarService.java
@@ -7,10 +7,10 @@ import java.util.Random;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.websoso.WSSServer.domain.Avatar;
-import org.websoso.WSSServer.domain.AvatarLine;
-import org.websoso.WSSServer.domain.AvatarProfile;
-import org.websoso.WSSServer.domain.AvatarProfileLine;
+import org.websoso.WSSServer.user.domain.Avatar;
+import org.websoso.WSSServer.user.domain.AvatarLine;
+import org.websoso.WSSServer.user.domain.AvatarProfile;
+import org.websoso.WSSServer.user.domain.AvatarProfileLine;
 import org.websoso.WSSServer.dto.avatar.AvatarProfileGetResponse;
 import org.websoso.WSSServer.dto.avatar.AvatarProfilesGetResponse;
 import org.websoso.WSSServer.repository.AvatarProfileRepository;
