+++ src/main/java/org/websoso/WSSServer/user/domain/AvatarProfile.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.domain;
+package org.websoso.WSSServer.user.domain;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
